Added Back-To-Top Button using JQuery
I've created a nice back-to-top button using JQuery, a nice feature to
have. It also suits my design as it there is a lot of scrolling, which
is very necessary to have. I might have to delay the overlay feature,
its a tough challenge, and I'm not sure how I'm going to implement it,
but we'll see how it goes.