Postponing future development on this project
I'm afraid I will have to postpone this project for now, what I'll do
now is get the basic prototype or "beta version" if you will, up and
running first, and find room for improvement later on in the process.
This will also mean that I will add thumnails for each completed mini
project, as well as a live preview link to the project, all you have to
do is click on the link and it will take you to the live preview, as
said before there will be no lightbox for it, just a pure hyper link.

Future improvements will include:
- A JQuery carousel, featuring one of each prorgamming language on the
nav. I will most likely be using a JQuery plugin.
- An overlay lightbox feature to show and explain the project as well as
a live preview button to click on.