Updates: Navigation, Footer, and Styling etc.
Big changes on this commit, I've removed the C# and iOS links because
those are projects I will develop in the future, not right now, I've
replaced with about and contact pages to explain what this projects is
and the purpose of it. I've also styled the footer, I'm trying to create
a footer navigation at the bottom. I've also made some small changes on
the styling in general, changed colors, shadows, sizes, and trying to
follow best practices.